Sanitize all parsed paths and wrap try/except over all file operations
This adds extensive sanitizing of all parsed content to attempt to clean corrupted path and filenames. It also wraps exception handling around operations like makedir, the reading of the backup file, and saving of all files. This should prevent one or two problem dirs/files from stopping the process because the script is handling the error. This addresses issue #8